// frontend/src/app/api/models/post.interface.ts
export interface Post {
  author: string;
  content: string;
  country: string | null;
  date_time: number;
  id: string;
  language: string;
  latitude: string | null;
  longitude: string | null;
  number_of_likes: number;
  number_of_shares: number;
  _id: string;
  _key: string;
  _rev: string;
}